+++ src/state/stage.js
@@ -1,6 +1,7 @@
 import { renderer } from "./../video/video.js";
 import * as  game from "./../game.js";
 import Camera2d from "./../camera/camera2d.js";
+import Color from "./../math/color.js";
 
 // a default camera instance to use across all stages
 var default_camera;
@@ -36,6 +37,40 @@ class Stage {
          * @memberof Stage
          */
         this.cameras = new Map();
+
+        /**
+         * The list of active lights in this stage.
+         * (Note: Canvas Renderering mode will only properly support one light per stage)
+         * @public
+         * @type {Map<Light2d>}
+         * @name lights
+         * @memberof Stage
+         * @see Light2d
+         * @see Stage.ambientLight
+         * @example
+         * // create a white spot light
+         * var whiteLight = new me.Light2d(0, 0, 140, "#fff", 0.7);
+         * // and add the light to this current stage
+         * this.lights.set("whiteLight", whiteLight);
+         * // set a dark ambient light
+         * this.ambientLight.parseCSS("#1117");
+         * // make the light follow the mouse
+         * me.input.registerPointerEvent("pointermove", me.game.viewport, (event) => {
+         *    whiteLight.centerOn(event.gameX, event.gameY);
+         * });
+         */
+        this.lights = new Map();
+
+        /**
+         * an ambient light that will be added to the stage rendering
+         * @public
+         * @type {Color}
+         * @name ambientLight
+         * @memberof Stage
+         * @default "#000000"
+         * @see Light2d
+         */
+        this.ambientLight = new Color(0, 0, 0, 0);
 
         /**
          * The given constructor options
@@ -92,7 +127,14 @@ class Stage {
         // update the camera/viewport
         // iterate through all cameras
         this.cameras.forEach(function(camera) {
-            if (camera.update(dt)) {
+            if (camera.update(dt) === true) {
+                isDirty = true;
+            };
+        });
+
+        // update all lights
+        this.lights.forEach((light) => {
+            if (light.update(dt) === true) {
                 isDirty = true;
             };
         });
@@ -110,9 +152,32 @@ class Stage {
      */
     draw(renderer) {
         // iterate through all cameras
-        this.cameras.forEach(function(camera) {
+        this.cameras.forEach((camera) => {
             // render the root container
             camera.draw(renderer, game.world);
+
+            // render the ambient light
+            if (this.ambientLight.alpha !== 0) {
+                renderer.save();
+                // iterate through all lights
+                this.lights.forEach((light) => {
+                    // cut out all lights visible areas
+                    renderer.setMask(light.getVisibleArea(), true);
+                });
+                // fill the screen with the ambient color
+                renderer.setColor(this.ambientLight);
+                renderer.fillRect(0, 0, camera.width, camera.height);
+                // clear all masks
+                renderer.clearMask();
+                renderer.restore();
+            }
+
+            // render all lights
+            this.lights.forEach((light) => {
+                light.preDraw(renderer, game.world);
+                light.draw(renderer, game.world);
+                light.postDraw(renderer, game.world);
+            });
         });
     }
 
@@ -123,6 +188,11 @@ class Stage {
     destroy() {
         // clear all cameras
         this.cameras.clear();
+        // clear all lights
+        this.lights.forEach((light) => {
+            light.destroy();
+        });
+        this.lights.clear();
         // notify the object
         this.onDestroyEvent.apply(this, arguments);
     }